separate html and javascript code
the function hide was in the header.html file enclosed by script tags. it has
been moved to the theme.js file.

to make it run it was necessary to add a entry point to the WebpackThemeBundle
instance in webpack.py (render the js file) and include the resulting js file
manually in footer.html (to be loaded in the browser).