index.html moved the divider style to the introduced articles
the divider divs and also the spacer div are not more necessary. it is possible
to apply the border-bottom to the article itself.

side effect is, that we have also a border-bottom for the last article. this is
intended. with that we have a better visualization that the "more" button belongs
to the whole list of records and not only for the last one