record_landing_page redesign access right visualisation
The green color of the access right badge could lead one to the false conclusion
that the record has a green open access. There exists a lot of different open
access ways like gold, bronze, green.

Tt may be possible to colorize the access right badge according to the real open
access way, but this would mean, that it has to be a field added to the metadata
and it has to be also put a value there. This is a little to much effort to only
have a colorized badge and more important it has to be decided on another level
as the design of the record view.

Therefore it is easier to remove the green color and colorize the text with the
orange open access color.

Next problem is then that there is a badge without a colorized box around it,
which leads to the conclusion that it has to be moved to another position on the
record view.

Finally it is now where the contact send button was.