record_landing_page override the doi.html with a local one
The reason was mainly to remove the DOI header above the doi badge. It was a
redundant information.

Further it was easier to apply the tugraz theme font-family and the font-size to
the doi-modal.

Further it was also possible to change the code to a correct html. dt and dd
should not be a child group of a div. The correct parent should be dl. [1]

[1] https://developer.mozilla.org/en-US/docs/Web/HTML/Element/dl