views make_dict_like rewrite the function to make it more general
the function was designed to fulfil a special need, the glue between
the response of record.access_right and the vocabolary_title function.
vocabolary_title has been designed to get a elasticsearch object, and
on the frontpage we get the data directly from the database and it has
thefore not the same structure as the elasticsearch return value.

now the function could be used also for other record fields if it
would be necessary. this generalization was achieved by introducing a
second parameter as the key for the value in the dict.