RDMRecordResultsListItem bugfix nested a element not possible
due to the added orcid link there is a nested a element. the whole
search result is a clickable link, this is is problem.

now there is the title and the description a clickable link.

this has also produced a followup error. the description margin bottom
is not more working. there was a display: block to add.